Also show mic with picovoice (#680)
PicoVoice requires a small time to load the model. Therefore the goal is active before PV is actually working. Therefore the mic will show, before it is working. But this is delay is not that much. But when we experience this is the case, we need to adapt to this.